Began Updating Patterns with 3 Color

+++ Responder/Responder/src/Communication/Command.h
@@ -5,13 +5,20 @@
 #ifndef COMMAND_H
 #define COMMAND_H
 
+#include <FastLED.h>
+
+/*
+ * Struct to pass Commands between Serial Reader class and
+ * Operator class
+*/
+
 typedef struct Command {
     enum {COLOR, STATIC_P, MUSIC_P, NONE};
     int type;
     int value;
-    unsigned short r;
-    unsigned short g;
-    unsigned short b;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
 } Command_t;
 
 #endif //COMMAND_H
